% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ilo.r
\name{check_ilo}
\alias{check_ilo}
\title{helper to calculate ilo indicator (volume and ration)}
\usage{
check_ilo(df, ...)
}
\arguments{
\item{df}{microdataset.}

\item{...}{Specification of variables to get.}
}
\value{
data frame with check columns
}
\description{
faster method to recode, order, add label of categorical variables
}
\details{
Support for recoding variable, check \code{?car::Recode}, for labelling variable
}
\section{Specification}{

others functions allow advanced manipulation.
\itemize{
 \item \code{scope}: check set as TRUE if indicator, sex, classif1, classif2 are ok, FALSE else,
 \item \code{version}: check set as TRUE if collection indicator configuration is ok , FALSE else,
 \item \code{indicator}: check set as TRUE if collection indicator, FALSE else,
}
}

\examples{
## Not run:

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{preprocessing}
