% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ilo.r
\name{compare_ilo}
\alias{compare_ilo}
\title{R tools to compare ilo datasets with a self made datasets}
\usage{
compare_ilo(df, ...)
}
\arguments{
\item{df}{ilo tbl data frame to manipulate.}

\item{...}{Specification of columns and pattern.}
}
\value{
a tbl data frame
}
\description{
wraper
}
\section{Specification}{

These functions allow you to select database variables based on their names.
\itemize{
 \item \code{collection}: select collection, characters verctor, default is 'YI', see ilo$segment, also available 'all',
 \item \code{freq}: select freqency, characters verctor, default is 'A', see ilo$segment, also available 'all',
 \item \code{ref_area}: select ref_area, characters verctor, default is 'all',
 \item \code{indicator}: used for searching character string in indicator variable,
 \item \code{source}: used for searching character string in source variable,
 \item \code{source.ty}: used for searching character string with the 2 first characters of source variable,
 \item \code{time}: used for searching character string in time variable,
 \item \code{obs_status}: used for searching character string in obs_status variable,
 \item \code{note_indicator}: used for searching character string in note_indicator variable,
 \item \code{note_source}: used for searching character string in note_source variable,
 \item \code{info}: used for searching character string in info administrative variable, if not call will not appear,
 \item \code{web}: base on backoffice data and meta config, return value that are publish on the website,
}
To learn more about ilo, start with the online vignettes:
\code{compare_ilo()}
}

\examples{
res <- get_ilo(ref_area = 'AFG')

res \%>\% filter_ilo(note_source = 'T3')
}
