% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ilo.r
\name{get_ilo}
\alias{get_ilo}
\title{R tools to access internal data on oracle / ilostat}
\usage{
get_ilo(...)
}
\arguments{
\item{...}{Specification of selection and/or queries.}
}
\value{
a tbl data frame
}
\description{
R tools to access internal data on oracle / ilostat
}
\section{Specification}{

These functions allow you to select database variables based on their names.
\itemize{
 \item \code{collection}: select collection, characters verctor, default is 'YI', see ilo$segment, also available 'all',
 \item \code{freq}: select freqency, characters verctor, default is 'A', see ilo$segment, also available 'all',
 \item \code{ref_area}: select ref_area, characters verctor, default is 'all',
 \item \code{indicator}: used for searching character string in indicator variable,
 \item \code{source}: used for searching character string in source variable,
 \item \code{sex}: used for searching character string in sex variable,
 \item \code{classif1}: used for searching character string in classif1 variable,
 \item \code{classif2}: used for searching character string in classif2 variable,
 \item \code{time}: used for searching character string in time variable,
 \item \code{obs_status}: used for searching character string in obs_status variable,
 \item \code{note_classif}: used for searching character string in note_classif variable,
 \item \code{note_indicator}: used for searching character string in note_indicator variable,
 \item \code{note_source}: used for searching character string in note_source variable,
 \item \code{info}: used for searching character string in info administrative variable, if not call will not appear,
 \item \code{web}: base on backoffice data and meta config, return value that are publish on the website,
}

others functions allow advanced manipulation.
\itemize{
 \item \code{label}: return label and allow to work with label (not capital sensitive) for the variables mentionned above,
 \item \code{timefrom}: filter time >= chracter string of year, ie year >= 2000 is timefrom = '2000',
 \item \code{timeto}: filter time <= chracter string of year, ie year <= 2005 is timeto = '2005',
 \item \code{style}: to be completed, default 'none' also available 'ts',
 \item \code{query}: used for inserting dplyr manipulation code at inside each collection/freq/ref_area segments,
 \item \code{add}: R object that should be add if needed for the parallel process of the query,
 \item \code{package}: R package that should be add if needed for the parallel process of the query,
}
To learn more about ilo, start with the online vignettes:
\code{help_ilo()}
}

\examples{
\dontrun{
init_ilo()


### quarterly time serie of female unemployed as from 2000 in united states

get_ilo(collection = 'STI', 
	freq = 'Q', ref_area = 'USA', 
	indicator = 'UNE_TUNE_SEX_AGE_NB', 
	timefrom = '2000', sex = 'F', classif1 = 'AGGREGATE_TOTAL', 
	query = 'select(ref_area:obs_value) \%>\% spread(time, obs_value)', package = 'tidyr')

### Annual data of Afganistan from Yearly Collection

get_ilo(ref_area = 'AFG')

### Annual collected data  from Yearly Collection

get_ilo(info = 'COL')

### Working with label

get_ilo(label, ref_area = 'AFG', source = 'Census', indicator = 'population')
}
}
