% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilo.r
\name{ilo}
\alias{ilo}
\alias{rnm}
\title{R tools to access data from [ILO Statistics department]{http://www.ilo.org/ilostat/)}}
\usage{
rnm(x)
}
\description{
ilo package provides a flexible grammar of ilo data manipulation. Associate with dplyr, tidyr and others packages developped by Hadley Wickham & Co, it provide a faster way to work labour statistics from the ilo.
}
\details{
It has four main goals:

\itemize{
\item Identify and get data from ilo database.
\item Support reproductible analysis.
\item View and map data or summary for analytical purpose (in progress).
\item Provide fast performance for in-memory data by using dplyr and parallel computation.
}

ilo package always need to be initialise: code{init_ilo()} in order to prepare parallel computation and setting codelist

To learn more about ilo, start with the online vignettes:
\code{help_ilo()}
}
\examples{
\dontrun{
init_ilo()
get_ilo(ref_area = 'AFG')
}
}
