% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ilo.r, R/init_ilo.r
\name{init_ilo}
\alias{init_ilo}
\alias{filter_ilo}
\alias{init_ilo}
\title{R tools to manipulate ilo datasets code / label}
\usage{
filter_ilo(df, ...)

init_ilo(...)
}
\arguments{
\item{df}{ilo tbl data frame to manipulate.}

\item{...}{Specification of columns and pattern.}

\item{...}{Specification for extenal update.}
}
\value{
a tbl data frame

ilo an R environment that contains cluster and codelist
}
\description{
wraper for filter(string::str_detect(string, pattern))

init ilo packages by loading codelist and cluster for parallel computation
}
\section{Specification}{

To learn more about ilo, start with the online vignettes:
\code{help_ilo()}


others functions allow advanced manipulation.
\itemize{
 \item \code{update}: update database for external users,
}
To learn more about ilo, start with the online vignettes:
\code{help_ilo()}
}

\examples{
\dontrun{
res <- get_ilo(ref_area = 'AFG')

res \%>\% filter_ilo(note_source = 'T3')
}
\dontrun{
init_ilo()

ilo
ilo \%>\% names
ilo$lastupdate
ilo$code \%>\% names
ilo$code$cl_country
}
}
