% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ilo.r
\name{save_ilo}
\alias{save_ilo}
\title{R tools to manipulate ilo datasets code / label}
\usage{
save_ilo(...)
}
\arguments{
\item{...}{Specification of savinf format.}

\item{df}{ilo tbl data frame}
}
\value{
external file open in temp workspace
}
\description{
save and open easily your dataset in an xls, csv, dta files format
}
\section{Specification}{

specification allow for saving.
\itemize{
 \item \code{format}: specify format to save, if format is not specify, xlsx will be used.
}
setting format .
\itemize{
 \item \code{xls}: allow to save multiple dataset on different .xlsx sheet (could be also rename),
 \item \code{csv}: allow one dataset to be save in .csv, is also use for ilo dataset for uploading data,
 \item \code{dta}: allow one dataset to be save in .dta,
 \item \code{del}: store dataset in deletion ilo format .csv,
 \item \code{rev}: store dataset in del ilo format as well as upload ilo format in .xlsx,
 \item \code{path}: if set will store file in your path directory, see example.
 \item \code{open}: if path is set, you could just save file and not open it (set as FALSE), default = TRUE.
}
To learn more about ilo, start with the online vignettes:
\code{help_ilo()}
}

\examples{
\dontrun{
### quarterly female unemployed as from 2000 in united states

res <-  get_ilo(collection = 'STI', 
	freq = 'Q', ref_area = 'USA', 
	indicator = 'UNE_TUNE_SEX_AGE_NB', 
	timefrom = '2000', sex = 'F', classif1 = 'AGGREGATE_TOTAL')
save_ilo(res = 'MyNewTab')
res \%>\% save_ilo(format = 'csv')
res \%>\% save_ilo(format = 'del')
res \%>\% save_ilo(format = 'rev')

### save multi shhet xls file
sum = res \%>\% count(time) 

save_ilo(sum = 'summary' , res = 'raw')

### save file on disk

save_ilo(res , path = 'D:/test.csv')

}
}
