% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_ilo.r
\name{switch_ilo}
\alias{switch_ilo}
\title{R tools to manipulate ilo datasets code / label}
\usage{
switch_ilo(df, ...)
}
\arguments{
\item{df}{ilo tbl data frame to manipulate.}

\item{...}{Specification of variables to get.}
}
\value{
a tbl data frame
}
\description{
ilo package provides a flexible grammar of ilo data manipulation. Associate with dplyr, tidyr and others packages developped by Hadley Wickham & Co, it provide a faster way to work labour statistics from the ilo.
}
\section{Specification}{

others functions allow advanced manipulation.
\itemize{
 \item \code{version}: return data frame with classif version code,
 \item \code{keep}: keep code when label is ask,
}
To learn more about ilo, start with the online vignettes:
}

\examples{
### quarterly female unemployed as from 2000 in united states

res <-  get_ilo(collection = 'STI', 
	freq = 'Q', ref_area = 'USA', 
	indicator = 'UNE_TUNE_SEX_AGE_NB', 
	timefrom = '2000', sex = 'F', classif1 = 'AGGREGATE_TOTAL')
res \%>\% switch_ilo(keep)

### get version code

res \%>\% switch_ilo(version)

### get label of collection and contry only (don't keep code)

res \%>\% switch_ilo(collection, ref_area)

}
